Bootstrap: memory changes
Changed the bootstrapper to map only available mmap entries. This should
reduce the size of the HHDM ideally stop itself from mapping to much
memory. Additional checks should be implemented.

+++ bootstrap/src/c/include/arctan.h
@@ -28,7 +28,7 @@
 #ifndef ARC_ARCTAN_H
 #define ARC_ARCTAN_H
 
-#define ARC_HHDM_VADDR   0xFFFFC00000000000 // 192 TiB
+#define ARC_HHDM_VADDR   (uint64_t)0xFFFFC00000000000 // 192 TiB
 #define ARC_PHYS_TO_HHDM(physical) ((uintptr_t)(physical) + (uintptr_t)ARC_HHDM_VADDR)
 #define ARC_HHDM_TO_PHYS(hhdm) ((uintptr_t)(hhdm) - (uintptr_t)ARC_HHDM_VADDR)
